Added txt file of python console and support dev file
Cake_walk was the count and sort and filter.  Very small script.
The path.txt is the path that I took to finally figure out the hash
problem.  It shows the repeatedly-failed and stubborn brute-force
attempt.  It also shows how I worked out the base 37 method.  The part
where I slept is less apparent.